feat: refactor route decomposition, MVC isolation & ownership middleware (PR #14)
